requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.6.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
Pillow==8.0.1
psycopg2==2.8.6
pypng==0.0.20
PyQRCode==1.2.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
